feat: ensemble trained and benchmarked — beats XGBoost baseline
Benchmark results (walk-forward OOS):
- Ensemble AUC: 0.8369 vs XGBoost 0.8210 (+0.0159)
- Full-sample: Ensemble AUC 0.9210 vs baseline 0.8759 (+0.0451)
- On 17 disagreements: ensemble correct 71% vs baseline 29%
- Ensemble weights: XGB 31.8%, RF 34.3%, ET 33.8%

Also: ml_strategy.py updated with ensemble_mode config flag,
report generator, ML improvement research notes

+++ compass/ensemble_signal_model.py
@@ -269,7 +269,7 @@ class EnsembleSignalModel:
             logger.info("Training ensemble on %d samples...", len(features_df))
 
             self.feature_names = list(features_df.columns)
-            X = sanitize_features(features_df.values)
+            X = sanitize_features(features_df.values.astype(np.float64))
             y = labels
 
             # 1. Hold out 20% test set
@@ -442,7 +442,7 @@ class EnsembleSignalModel:
             return np.ones(len(features_df)) * 0.5
 
         try:
-            X = sanitize_features(features_df[self.feature_names].values)
+            X = sanitize_features(features_df[self.feature_names].values.astype(np.float64))
             self._check_feature_distribution(X)
             return self._weighted_predict_proba(X)
